modularize quick pairing tests
