+++ apps/server/src/routes/game/quick_pairing_game/mod.rs
@@ -37,3 +37,6 @@ pub fn docs(op: TransformOperation) -> TransformOperation {
         .response::<200, Json<GameId>>()
         .response::<400, Json<GenericError>>()
 }
+
+#[cfg(test)]
+mod tests;
